refactor(gex): modularize JS and CSS, add FastAPI server
JavaScript:
- Split main.js into state.js, ui.js, chart.js, simulation.js, events.js
- main.js now orchestration only (~170 lines vs ~1580)
- Uses window.GEX namespace for cross-module communication

CSS:
- Split styles.css into 8 component files in css/ folder
- Uses @import for clean organization

Backend:
- Add server.py (FastAPI) for production use
- Update run.py with FastAPI fallback to stdlib
