+++ tools/gex-visualizer/css/base.css
@@ -1,0 +1,50 @@
+/* GEX Visualizer - Base Styles */
+/* CSS variables, reset, body */
+
+:root {
+    --bg-dark: #0a0d10;
+    --panel-bg: #12171d;
+    --card-bg: #1a2029;
+    --text-main: #e8eaed;
+    --text-dim: #8b9bb4;
+    --accent-blue: #2962ff;
+    --accent-cyan: #00e5ff;
+    --accent-red: #ff3d57;
+    --accent-green: #00c853;
+    --accent-orange: #ff9100;
+    --grid-line: #2a323d;
+    --warning-yellow: #ffc107;
+    --purple: #9c27b0;
+}
+
+* { box-sizing: border-box; }
+
+body {
+    margin: 0;
+    font-family: 'Segoe UI', 'Roboto', sans-serif;
+    background-color: var(--bg-dark);
+    color: var(--text-main);
+    overflow: hidden;
+    height: 100vh;
+    display: flex;
+    flex-direction: column;
+}
+
+/* Main Layout */
+.dashboard {
+    display: grid;
+    grid-template-columns: var(--sidebar-width, 280px) 1fr 1fr;
+    flex: 1;
+    overflow: hidden;
+}
+
+/* Animations */
+@keyframes pulse {
+    0%, 100% { opacity: 1; transform: scale(1); }
+    50% { opacity: 0.5; transform: scale(1.2); }
+}
+
+@keyframes glow {
+    0%, 100% { box-shadow: 0 0 5px rgba(255,61,87,0.5); }
+    50% { box-shadow: 0 0 20px rgba(255,61,87,0.8); }
+}
