+++ tools/gex-visualizer/css/cards.css
@@ -1,0 +1,143 @@
+/* GEX Visualizer - Card Styles */
+/* Zero gamma, persistence, performance, lead-lag cards */
+
+/* Zero Gamma Level */
+.zero-gamma-card {
+    background: linear-gradient(135deg, rgba(156,39,176,0.2) 0%, rgba(41,98,255,0.2) 100%);
+    border: 1px solid var(--purple);
+    border-radius: 8px;
+    padding: 12px;
+}
+
+.zero-gamma-value {
+    font-size: 1.5rem;
+    font-weight: bold;
+    color: var(--purple);
+    text-align: center;
+    margin: 8px 0;
+}
+
+.zero-gamma-label {
+    font-size: 0.65rem;
+    color: var(--text-dim);
+    text-align: center;
+}
+
+.gamma-flip-indicator {
+    display: flex;
+    justify-content: space-between;
+    font-size: 0.6rem;
+    margin-top: 8px;
+    padding-top: 8px;
+    border-top: 1px solid var(--grid-line);
+}
+
+.flip-zone {
+    padding: 3px 8px;
+    border-radius: 3px;
+    font-weight: 600;
+}
+
+.flip-below { background: rgba(0,200,83,0.2); color: var(--accent-green); }
+.flip-above { background: rgba(255,61,87,0.2); color: var(--accent-red); }
+
+/* Regime Persistence */
+.persistence-card {
+    background: var(--card-bg);
+    border: 1px solid var(--grid-line);
+    border-radius: 8px;
+    padding: 12px;
+}
+
+.persistence-meter {
+    display: flex;
+    gap: 3px;
+    margin: 8px 0;
+}
+
+.persistence-bar {
+    flex: 1;
+    height: 20px;
+    background: var(--grid-line);
+    border-radius: 2px;
+    transition: background 0.3s;
+}
+
+.persistence-bar.active { background: var(--accent-cyan); }
+.persistence-bar.active.negative { background: var(--accent-red); }
+.persistence-bar.active.positive { background: #2ed573; }
+
+.persistence-stats {
+    display: flex;
+    justify-content: space-between;
+    font-size: 0.6rem;
+    color: var(--text-dim);
+}
+
+/* Performance Metrics */
+.perf-card {
+    background: var(--card-bg);
+    border: 1px solid var(--grid-line);
+    border-radius: 8px;
+    padding: 12px;
+}
+
+.perf-row {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    padding: 6px 0;
+    border-bottom: 1px solid var(--grid-line);
+}
+
+.perf-row:last-child { border-bottom: none; }
+
+.perf-label { font-size: 0.7rem; color: var(--text-dim); }
+.perf-value { font-size: 0.8rem; font-weight: bold; }
+.perf-positive { color: var(--accent-green); }
+
+.perf-badge {
+    background: var(--accent-green);
+    color: black;
+    padding: 2px 6px;
+    border-radius: 3px;
+    font-size: 0.6rem;
+    font-weight: bold;
+}
+
+/* Lead-Lag Indicator */
+.lead-lag-card {
+    background: var(--card-bg);
+    border: 1px solid var(--grid-line);
+    border-radius: 8px;
+    padding: 10px;
+}
+
+.lead-lag-visual {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    gap: 8px;
+    padding: 8px 0;
+}
+
+.lead-lag-symbol {
+    padding: 4px 10px;
+    border-radius: 4px;
+    font-size: 0.75rem;
+    font-weight: bold;
+}
+
+.uvxy-badge { background: rgba(255,61,87,0.2); color: var(--accent-red); }
+.spy-badge { background: rgba(0,229,255,0.2); color: var(--accent-cyan); }
+
+.lead-arrow {
+    color: var(--accent-orange);
+    font-size: 1.2rem;
+}
+
+.lead-lag-stats {
+    text-align: center;
+    font-size: 0.6rem;
+    color: var(--text-dim);
+}
