+++ tools/gex-visualizer/css/charts.css
@@ -1,0 +1,195 @@
+/* GEX Visualizer - Chart Styles */
+/* Chart panels, viz container, axes, SVG elements */
+
+/* Chart Panels */
+.chart-panel {
+    position: relative;
+    display: flex;
+    flex-direction: column;
+    border-right: 1px solid var(--grid-line);
+    cursor: grab;
+}
+
+.panel-header {
+    padding: 12px 15px;
+    background: var(--panel-bg);
+    border-bottom: 1px solid var(--grid-line);
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+}
+
+.panel-title {
+    font-weight: 600;
+    font-size: 0.85rem;
+    color: var(--accent-cyan);
+}
+
+.panel-subtitle {
+    font-size: 0.65rem;
+    color: var(--text-dim);
+    margin-top: 2px;
+}
+
+.panel-status {
+    text-align: right;
+}
+
+.status-label {
+    font-size: 0.6rem;
+    color: var(--text-dim);
+    text-transform: uppercase;
+}
+
+.status-value {
+    font-size: 0.85rem;
+    font-weight: bold;
+    color: var(--text-dim);
+}
+
+/* Viz Container */
+.viz-container {
+    flex: 1;
+    display: flex;
+    position: relative;
+    background: linear-gradient(180deg, transparent 0%, rgba(0,0,0,0.3) 100%);
+}
+
+/* Y-Axis */
+.y-axis {
+    width: 45px;
+    display: flex;
+    flex-direction: column;
+    justify-content: space-between;
+    padding: 8px 4px;
+    font-size: 0.6rem;
+    color: var(--text-dim);
+    text-align: right;
+    background: rgba(0,0,0,0.2);
+    position: relative;
+    cursor: ns-resize;
+    transition: background 0.2s;
+}
+
+.y-axis:hover {
+    background: rgba(0,229,255,0.15);
+}
+
+.y-axis .y-label {
+    transition: color 0.2s;
+}
+
+.y-axis:hover .y-label {
+    color: var(--accent-cyan);
+}
+
+.y-axis::after {
+    content: attr(data-scale);
+    position: absolute;
+    bottom: 2px;
+    right: 4px;
+    font-size: 0.5rem;
+    color: var(--text-dim);
+    opacity: 0;
+    transition: opacity 0.2s;
+}
+
+.y-axis:hover::after {
+    opacity: 1;
+}
+
+/* SVG Container */
+.svg-wrapper {
+    flex: 1;
+    position: relative;
+    display: flex;
+    flex-direction: column;
+}
+
+.svg-container {
+    flex: 1;
+    position: relative;
+}
+
+.svg-container svg { width: 100%; height: 100%; }
+
+/* X-Axis */
+.x-axis {
+    height: 24px;
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    padding: 0 8px;
+    font-size: 0.55rem;
+    color: var(--text-dim);
+    background: rgba(0,0,0,0.2);
+    cursor: ew-resize;
+    position: relative;
+    transition: background 0.2s;
+}
+
+.x-axis:hover {
+    background: rgba(0,229,255,0.1);
+}
+
+.x-axis:hover .x-label {
+    color: var(--accent-cyan);
+}
+
+.x-axis-center {
+    position: absolute;
+    left: 50%;
+    transform: translateX(-50%);
+    color: var(--text-dim);
+    font-weight: bold;
+}
+
+.x-axis::after {
+    content: attr(data-scale);
+    position: absolute;
+    bottom: 2px;
+    left: 50%;
+    transform: translateX(-50%);
+    font-size: 0.45rem;
+    color: var(--accent-orange);
+    opacity: 0;
+    transition: opacity 0.2s;
+}
+
+.x-axis:hover::after {
+    opacity: 1;
+}
+
+/* SVG Elements */
+svg { width: 100%; height: 100%; }
+.bar { transition: all 0.15s ease; opacity: 0.85; }
+.bar:hover { opacity: 1; }
+.center-line { stroke: var(--text-dim); stroke-width: 1; stroke-dasharray: 3,3; opacity: 0.4; }
+.price-line { stroke: var(--warning-yellow); stroke-width: 2; }
+.zero-gamma-line { stroke: var(--purple); stroke-width: 2; stroke-dasharray: 5,3; }
+.price-marker { fill: var(--warning-yellow); font-size: 9px; font-weight: bold; }
+.zero-marker { fill: var(--purple); font-size: 9px; font-weight: bold; }
+
+/* Data Overlay */
+.data-overlay {
+    position: absolute;
+    bottom: 38px;
+    left: 50px;
+    background: rgba(0,0,0,0.85);
+    padding: 10px 12px;
+    border: 1px solid var(--grid-line);
+    border-radius: 6px;
+    backdrop-filter: blur(4px);
+    z-index: 10;
+}
+
+.data-row {
+    display: flex;
+    justify-content: space-between;
+    gap: 15px;
+    font-size: 0.7rem;
+    margin-bottom: 3px;
+}
+
+.data-row:last-child { margin-bottom: 0; }
+.data-val { font-weight: bold; }
