+++ tools/gex-visualizer/css/controls.css
@@ -1,0 +1,262 @@
+/* GEX Visualizer - Controls Styles */
+/* Simulation, media controls, timeline, sliders */
+
+/* Simulation Controls */
+.sim-controls {
+    background: var(--card-bg);
+    border-radius: 8px;
+    padding: 12px;
+    border: 1px solid var(--grid-line);
+}
+
+.btn-group {
+    display: flex;
+    gap: 8px;
+    margin-bottom: 10px;
+}
+
+.sim-btn {
+    flex: 1;
+    background: linear-gradient(135deg, var(--accent-blue) 0%, #1e4bd1 100%);
+    color: white;
+    border: none;
+    padding: 10px;
+    border-radius: 6px;
+    cursor: pointer;
+    font-family: inherit;
+    font-weight: 600;
+    font-size: 0.75rem;
+    transition: all 0.2s;
+}
+
+.sim-btn:hover { transform: translateY(-1px); box-shadow: 0 4px 12px rgba(41,98,255,0.4); }
+
+.sim-active {
+    background: linear-gradient(135deg, var(--accent-red) 0%, #d32f2f 100%);
+    animation: glow 1.5s infinite;
+}
+
+.reset-btn {
+    background: var(--card-bg);
+    border: 1px solid var(--grid-line);
+    flex: 0.4;
+}
+
+.reset-btn:hover { background: #2a323d; transform: none; box-shadow: none; }
+
+/* Media Controls */
+.media-controls {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    gap: 6px;
+    margin-bottom: 10px;
+}
+
+.media-btn {
+    width: 28px;
+    height: 28px;
+    border-radius: 50%;
+    border: 1px solid var(--grid-line);
+    background: var(--card-bg);
+    color: var(--text-main);
+    cursor: pointer;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    font-size: 10px;
+    transition: all 0.2s;
+    padding: 0;
+}
+
+.media-btn:hover {
+    background: var(--grid-line);
+    border-color: var(--accent-cyan);
+}
+
+.media-btn.play-btn {
+    width: 38px;
+    height: 38px;
+    background: linear-gradient(135deg, var(--accent-blue) 0%, #1e4bd1 100%);
+    border: none;
+    font-size: 14px;
+}
+
+.media-btn.play-btn:hover {
+    transform: scale(1.05);
+    box-shadow: 0 4px 12px rgba(41,98,255,0.5);
+}
+
+.media-btn.play-btn.playing {
+    background: linear-gradient(135deg, var(--accent-red) 0%, #d32f2f 100%);
+}
+
+.speed-control {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    gap: 4px;
+    margin-top: 8px;
+}
+
+.speed-btn {
+    padding: 4px 10px;
+    border-radius: 4px;
+    border: 1px solid var(--grid-line);
+    background: var(--card-bg);
+    color: var(--text-dim);
+    font-size: 0.65rem;
+    cursor: pointer;
+    transition: all 0.2s;
+}
+
+.speed-btn:hover, .speed-btn.active {
+    background: var(--accent-blue);
+    border-color: var(--accent-blue);
+    color: white;
+}
+
+/* Timeline Scrubber */
+.timeline-scrubber {
+    position: relative;
+    margin: 12px 0 8px;
+}
+
+.timeline-track {
+    height: 8px;
+    background: var(--bg-dark);
+    border-radius: 4px;
+    position: relative;
+    cursor: pointer;
+}
+
+.timeline-progress {
+    height: 100%;
+    background: linear-gradient(90deg, var(--accent-blue), var(--accent-cyan));
+    border-radius: 4px;
+    width: 0%;
+    transition: width 0.05s linear;
+    pointer-events: none;
+}
+
+.timeline-handle {
+    position: absolute;
+    top: 50%;
+    transform: translate(-50%, -50%);
+    width: 16px;
+    height: 16px;
+    background: var(--accent-cyan);
+    border-radius: 50%;
+    cursor: grab;
+    box-shadow: 0 2px 8px rgba(0,229,255,0.5);
+    transition: transform 0.1s;
+}
+
+.timeline-handle:hover {
+    transform: translate(-50%, -50%) scale(1.2);
+}
+
+.timeline-handle:active {
+    cursor: grabbing;
+}
+
+.timeline-markers {
+    display: flex;
+    justify-content: space-between;
+    margin-top: 6px;
+    padding: 0 2px;
+}
+
+.timeline-marker {
+    font-size: 0.55rem;
+    color: var(--text-dim);
+    cursor: pointer;
+    padding: 2px 4px;
+    border-radius: 3px;
+    transition: all 0.2s;
+}
+
+.timeline-marker:hover {
+    color: var(--accent-cyan);
+    background: rgba(0,229,255,0.1);
+}
+
+.timeline-marker.active {
+    color: var(--accent-cyan);
+    font-weight: bold;
+}
+
+/* Current Event Display */
+.current-event {
+    background: var(--bg-dark);
+    border-radius: 6px;
+    padding: 10px;
+    margin-top: 8px;
+    text-align: center;
+}
+
+.event-date {
+    font-size: 1rem;
+    font-weight: bold;
+    color: var(--accent-cyan);
+}
+
+.event-label {
+    font-size: 0.7rem;
+    color: var(--text-dim);
+    margin-top: 4px;
+}
+
+.event-index {
+    font-size: 0.6rem;
+    color: var(--text-dim);
+    margin-top: 6px;
+    opacity: 0.7;
+}
+
+.event-note {
+    font-size: 0.55rem;
+    color: var(--accent-orange);
+    margin-top: 8px;
+    padding-top: 6px;
+    border-top: 1px solid var(--grid-line);
+    font-style: italic;
+}
+
+/* Sliders */
+.control-group {
+    display: flex;
+    flex-direction: column;
+    gap: 6px;
+}
+
+.control-label {
+    display: flex;
+    justify-content: space-between;
+    font-size: 0.7rem;
+    color: var(--text-dim);
+}
+
+.control-value {
+    color: var(--accent-cyan);
+    font-weight: 600;
+}
+
+input[type="range"] {
+    width: 100%;
+    height: 4px;
+    background: var(--grid-line);
+    border-radius: 2px;
+    -webkit-appearance: none;
+    outline: none;
+}
+
+input[type="range"]::-webkit-slider-thumb {
+    -webkit-appearance: none;
+    width: 14px;
+    height: 14px;
+    border-radius: 50%;
+    background: var(--accent-cyan);
+    cursor: pointer;
+    box-shadow: 0 2px 6px rgba(0,229,255,0.4);
+}
