+++ tools/gex-visualizer/css/header.css
@@ -1,0 +1,168 @@
+/* GEX Visualizer - Header Styles */
+/* Header, logo, metric cards, year badge, volatility warning */
+
+header {
+    padding: 12px 20px;
+    background: linear-gradient(180deg, var(--panel-bg) 0%, var(--bg-dark) 100%);
+    border-bottom: 1px solid var(--grid-line);
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+}
+
+.header-left {
+    display: flex;
+    align-items: center;
+    gap: 15px;
+}
+
+.logo {
+    display: flex;
+    align-items: center;
+    gap: 10px;
+}
+
+.logo-icon {
+    width: 32px;
+    height: 32px;
+    background: linear-gradient(135deg, var(--accent-red) 0%, var(--accent-orange) 100%);
+    border-radius: 6px;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    font-size: 18px;
+}
+
+h1 {
+    margin: 0;
+    font-size: 1.1rem;
+    font-weight: 600;
+    letter-spacing: 0.5px;
+}
+
+.badge {
+    background: var(--accent-blue);
+    padding: 3px 8px;
+    border-radius: 4px;
+    font-size: 0.65rem;
+    font-weight: 600;
+    letter-spacing: 0.5px;
+}
+
+.header-center {
+    display: flex;
+    align-items: center;
+    gap: 20px;
+}
+
+.metric-card {
+    background: var(--card-bg);
+    border: 1px solid var(--grid-line);
+    border-radius: 6px;
+    padding: 8px 14px;
+    min-width: 100px;
+    text-align: center;
+}
+
+.metric-label {
+    font-size: 0.6rem;
+    color: var(--text-dim);
+    text-transform: uppercase;
+    letter-spacing: 0.5px;
+    margin-bottom: 2px;
+}
+
+.metric-value {
+    font-size: 1.1rem;
+    font-weight: bold;
+}
+
+.header-right {
+    display: flex;
+    align-items: center;
+    gap: 15px;
+}
+
+.year-badge {
+    background: linear-gradient(135deg, var(--accent-red) 0%, var(--purple) 100%);
+    padding: 8px 16px;
+    border-radius: 6px;
+    font-size: 1.4rem;
+    font-weight: bold;
+    letter-spacing: 2px;
+    min-width: 80px;
+    text-align: center;
+    box-shadow: 0 2px 10px rgba(255, 61, 87, 0.3);
+}
+
+.year-badge.inactive {
+    background: var(--card-bg);
+    box-shadow: none;
+    color: var(--text-dim);
+}
+
+/* Volatility Warning Banner */
+.vol-warning {
+    background: var(--panel-bg);
+    border-bottom: 1px solid var(--grid-line);
+    padding: 8px 20px;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    gap: 30px;
+    font-size: 0.75rem;
+    transition: all 0.3s ease;
+}
+
+.vol-warning.active.negative {
+    background: linear-gradient(90deg, rgba(255,61,87,0.15) 0%, rgba(255,145,0,0.15) 100%);
+    border-bottom-color: var(--accent-red);
+}
+
+.vol-warning.active.positive {
+    background: linear-gradient(90deg, rgba(0,229,255,0.1) 0%, rgba(46,213,115,0.1) 100%);
+    border-bottom-color: var(--accent-cyan);
+}
+
+.vol-warning.inactive {
+    opacity: 0.5;
+}
+
+.vol-stat {
+    display: flex;
+    align-items: center;
+    gap: 8px;
+}
+
+.vol-stat-value {
+    font-weight: bold;
+    color: var(--text-dim);
+    font-size: 0.9rem;
+    transition: color 0.3s;
+}
+
+.vol-warning.active.negative .vol-stat-value {
+    color: var(--accent-orange);
+}
+
+.vol-warning.active.positive .vol-stat-value {
+    color: var(--accent-cyan);
+}
+
+.vol-pulse {
+    width: 8px;
+    height: 8px;
+    background: var(--grid-line);
+    border-radius: 50%;
+    transition: background 0.3s;
+}
+
+.vol-warning.active.negative .vol-pulse {
+    background: var(--accent-red);
+    animation: pulse 1.5s infinite;
+}
+
+.vol-warning.active.positive .vol-pulse {
+    background: #2ed573;
+    animation: pulse 1.5s infinite;
+}
