+++ tools/gex-visualizer/css/sidebar.css
@@ -1,0 +1,235 @@
+/* GEX Visualizer - Sidebar Styles */
+/* Sidebar layout, section titles, data source controls */
+
+.sidebar {
+    background: var(--panel-bg);
+    border-right: 1px solid var(--grid-line);
+    padding: 15px;
+    display: flex;
+    flex-direction: column;
+    gap: 12px;
+    overflow-y: auto;
+    position: relative;
+    min-width: 220px;
+    max-width: 400px;
+}
+
+.sidebar-resize {
+    position: absolute;
+    top: 0;
+    right: 0;
+    width: 6px;
+    height: 100%;
+    cursor: ew-resize;
+    background: transparent;
+    transition: background 0.2s;
+    z-index: 10;
+}
+
+.sidebar-resize:hover,
+.sidebar-resize.dragging {
+    background: var(--accent-cyan);
+}
+
+.section-title {
+    font-size: 0.7rem;
+    color: var(--text-dim);
+    text-transform: uppercase;
+    letter-spacing: 1px;
+    margin-bottom: 5px;
+    display: flex;
+    align-items: center;
+    gap: 8px;
+}
+
+.section-title::after {
+    content: '';
+    flex: 1;
+    height: 1px;
+    background: var(--grid-line);
+}
+
+/* Data Source Controls */
+.data-source-controls {
+    background: var(--card-bg);
+    border-radius: 8px;
+    padding: 12px;
+    border: 1px solid var(--grid-line);
+    margin-bottom: 10px;
+}
+
+.data-mode-toggle {
+    display: flex;
+    gap: 4px;
+    margin-bottom: 10px;
+}
+
+.mode-btn {
+    flex: 1;
+    padding: 8px 12px;
+    border: 1px solid var(--grid-line);
+    background: var(--panel-bg);
+    color: var(--text-dim);
+    border-radius: 4px;
+    cursor: pointer;
+    font-size: 0.75rem;
+    font-weight: 600;
+    text-transform: uppercase;
+    transition: all 0.2s;
+}
+
+.mode-btn:hover {
+    background: var(--card-bg);
+    color: var(--text-main);
+}
+
+.mode-btn.active {
+    background: var(--accent-blue);
+    border-color: var(--accent-blue);
+    color: white;
+}
+
+.symbol-selectors {
+    display: flex;
+    flex-direction: column;
+    gap: 8px;
+}
+
+.selector-row select {
+    width: 100%;
+    padding: 8px 10px;
+    border: 1px solid var(--grid-line);
+    background: var(--panel-bg);
+    color: var(--text-main);
+    border-radius: 4px;
+    font-size: 0.8rem;
+    cursor: pointer;
+}
+
+.selector-row select:focus {
+    outline: none;
+    border-color: var(--accent-cyan);
+}
+
+.data-info {
+    font-size: 0.7rem;
+    color: var(--text-dim);
+    padding: 4px 0;
+}
+
+.data-info strong {
+    color: var(--accent-cyan);
+}
+
+/* Real Data Metrics Panel */
+.real-data-metrics {
+    margin-top: 10px;
+    padding-top: 10px;
+    border-top: 1px solid var(--grid-line);
+}
+
+.rdm-row {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    padding: 4px 0;
+    font-size: 0.7rem;
+}
+
+.rdm-label {
+    color: var(--text-dim);
+}
+
+.rdm-value {
+    font-weight: 600;
+    color: var(--text-main);
+}
+
+.rdm-value.quality-good { color: var(--accent-green); }
+.rdm-value.quality-fair { color: var(--warning-yellow); }
+.rdm-value.quality-poor { color: var(--accent-red); }
+.rdm-value.gamma-positive { color: var(--accent-cyan); }
+.rdm-value.gamma-negative { color: var(--accent-red); }
+
+/* Legend */
+.legend {
+    display: grid;
+    grid-template-columns: 1fr 1fr;
+    gap: 8px;
+}
+
+.legend-item {
+    display: flex;
+    align-items: center;
+    gap: 6px;
+    font-size: 0.65rem;
+    padding: 6px 8px;
+    background: var(--card-bg);
+    border-radius: 4px;
+}
+
+.legend-dot {
+    width: 10px;
+    height: 10px;
+    border-radius: 50%;
+}
+
+.dot-red { background: var(--accent-red); }
+.dot-cyan { background: var(--accent-cyan); }
+.dot-yellow { background: var(--warning-yellow); }
+.dot-purple { background: var(--purple); }
+
+/* Glossary / Terminology */
+.glossary {
+    background: var(--card-bg);
+    border: 1px solid var(--grid-line);
+    border-radius: 8px;
+    padding: 10px;
+}
+
+.glossary-item {
+    display: flex;
+    gap: 8px;
+    padding: 5px 0;
+    border-bottom: 1px solid var(--grid-line);
+    font-size: 0.6rem;
+    line-height: 1.4;
+}
+
+.glossary-item:last-child { border-bottom: none; }
+
+.glossary-term {
+    font-weight: bold;
+    color: var(--accent-cyan);
+    min-width: 50px;
+    flex-shrink: 0;
+}
+
+.glossary-def {
+    color: var(--text-dim);
+}
+
+/* Insight Box */
+.insight-box {
+    background: linear-gradient(135deg, rgba(41,98,255,0.1) 0%, rgba(0,229,255,0.1) 100%);
+    border: 1px solid var(--accent-blue);
+    border-radius: 8px;
+    padding: 12px;
+    margin-top: auto;
+}
+
+.insight-title {
+    font-size: 0.7rem;
+    font-weight: 600;
+    color: var(--accent-cyan);
+    margin-bottom: 6px;
+    display: flex;
+    align-items: center;
+    gap: 6px;
+}
+
+.insight-text {
+    font-size: 0.65rem;
+    line-height: 1.5;
+    color: var(--text-main);
+}
