+++ tools/gex-visualizer/index.html
@@ -392,7 +392,13 @@
     </div>
 </div>
 
+<!-- Module Loading Order: data-loader → state → ui → chart → simulation → events → main -->
 <script src="data-loader.js"></script>
+<script src="state.js"></script>
+<script src="ui.js"></script>
+<script src="chart.js"></script>
+<script src="simulation.js"></script>
+<script src="events.js"></script>
 <script src="main.js"></script>
 
 </body>
